--- features/lessons/src/main/java/com/cryart/sabbathschool/lessons/ui/lessons/components/spec/LessonItemSpec.kt
@@ -1,62 +1,0 @@
-/*
- * Copyright (c) 2022. Adventech <info@adventech.io>
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- */
-
-package com.cryart.sabbathschool.lessons.ui.lessons.components.spec
-
-import androidx.compose.runtime.Immutable
-import androidx.core.text.isDigitsOnly
-import app.ss.models.SSLesson
-import com.cryart.sabbathschool.core.misc.SSConstants
-import org.joda.time.format.DateTimeFormat
-
-@Immutable
-data class LessonItemSpec(
-    val index: String,
-    val title: String,
-    val date: String,
-)
-
-fun SSLesson.toSpec(): LessonItemSpec = LessonItemSpec(
-    index = if (id.isDigitsOnly()) {
-        "${id.toInt()}"
-    } else {
-        "•"
-    },
-    title = title,
-    date = dateDisplay(),
-)
-
-private fun SSLesson.dateDisplay(): String {
-    val startDateOut = DateTimeFormat.forPattern(SSConstants.SS_DATE_FORMAT_OUTPUT)
-        .print(
-            DateTimeFormat.forPattern(SSConstants.SS_DATE_FORMAT)
-                .parseLocalDate(start_date)
-        )
-
-    val endDateOut = DateTimeFormat.forPattern(SSConstants.SS_DATE_FORMAT_OUTPUT)
-        .print(
-            DateTimeFormat.forPattern(SSConstants.SS_DATE_FORMAT)
-                .parseLocalDate(end_date)
-        )
-
-    return "$startDateOut - $endDateOut".replaceFirstChar { it.uppercase() }
-}
